requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.3.2
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
requests==2.32.3
setuptools==69.5.1
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
backoff==2.2.1
